+++ src/main/resources/db/migration/R__create_provinces_stored_procedure.sql
@@ -1,7 +1,7 @@
 DROP PROCEDURE IF EXISTS province_find_all_by;
 DELIMITER //
 CREATE PROCEDURE IF NOT EXISTS province_find_all_by(
-    IN region_id INT,
+    IN region_id BIGINT,
     IN name VARCHAR(50),
     IN page INT,
     IN size INT
@@ -35,7 +35,7 @@ DELIMITER ;
 DROP PROCEDURE IF EXISTS province_find_all_by_total;
 DELIMITER //
 CREATE PROCEDURE IF NOT EXISTS province_find_all_by_total(
-    IN region_id INT,
+    IN region_id BIGINT,
     IN name VARCHAR(50)
 )
 BEGIN
